% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNA_Matrix_process_CLI.R,
%   R/CNA_Matrix_process_GUI.R
\name{CNA_Matrix_process_GUI}
\alias{CNA_Matrix_process_GUI}
\title{CNA_Matrix_process_GUI}
\usage{
CNA_Matrix_process_GUI(tar_gene)

CNA_Matrix_process_GUI(tar_gene)
}
\arguments{
\item{tar_gene}{}

\item{target_path}{The Dataset Filepath}
}
\value{
Return Processed CNA Datas(By tibble)

Return Processed CNA Datas(By tibble)
}
\description{
CNA_Matrix_process_GUI

CNA_Matrix_process_GUI
}
\examples{
test
}
