% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein_Matrix_process.R
\name{Protein_Matrix_process}
\alias{Protein_Matrix_process}
\title{Title}
\usage{
Protein_Matrix_process(tar_gene)
}
\arguments{
\item{tar_gene}{The Gene Name needed to be filtered}
}
\description{
Title
}
