% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein_Matrix_process_CLI.R
\name{Protein_Matrix_process_CLI}
\alias{Protein_Matrix_process_CLI}
\title{Protein_Matrix_process_CLI}
\usage{
Protein_Matrix_process_CLI(tar_gene, target_path)
}
\arguments{
\item{target_path}{}
}
\value{
Return Processed Protein Data(By Tibble)
}
\description{
Protein_Matrix_process_CLI
}
