% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protein_Matrix_process_GUI.R
\name{Protein_Matrix_process_GUI}
\alias{Protein_Matrix_process_GUI}
\title{Protein_Matrix_process_GUI}
\usage{
Protein_Matrix_process_GUI(tar_gene)
}
\arguments{
\item{tar_gene}{}
}
\value{
Return Processed Protein Data(By Tibble)
}
\description{
Protein_Matrix_process_GUI
}
