% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_Matrix_process_CLI.R
\name{RNA_Matrix_process_CLI}
\alias{RNA_Matrix_process_CLI}
\title{Title}
\usage{
RNA_Matrix_process_CLI(tar_gene, target_path)
}
\arguments{
\item{tar_gene}{The Gene Name needed to be filtered}

\item{target_path}{The Dataset Filepath}
}
\description{
Title
}
