% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_Matrix_process_GUI.R
\name{RNA_Matrix_process_GUI}
\alias{RNA_Matrix_process_GUI}
\title{Title}
\usage{
RNA_Matrix_process_GUI(tar_gene)
}
\arguments{
\item{tar_gene}{T}
}
\value{
Data Process Result
}
\description{
Title
}
